2022/11/17  20:53  完成了出纳的销售单和物流单的删除和新增

+++ oa-server/src/main/java/cn/edu/guet/controller/SaleContractController.java
@@ -1,12 +1,15 @@
 package cn.edu.guet.controller;
 
 import cn.edu.guet.bean.SysUser;
+import cn.edu.guet.bean.logisticsContract.LogisticsContract;
 import cn.edu.guet.bean.logisticsContract.LogisticsDetail;
 import cn.edu.guet.bean.sale.ListParm;
 import cn.edu.guet.bean.sale.SaleContract;
 import cn.edu.guet.http.HttpResult;
 import cn.edu.guet.http.ResultUtils;
+import cn.edu.guet.service.LogisticsContractService;
 import cn.edu.guet.service.SaleContractService;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import org.apache.coyote.Response;
 import org.springframework.web.bind.annotation.*;
@@ -28,6 +31,9 @@ public class SaleContractController {
 
     @Resource
     private SaleContractService saleContractService;
+
+    @Resource
+    private LogisticsContractService logisticsContractService;
 
     /**
      * 获取归档位1的分页销售单数据
@@ -74,6 +80,19 @@ public class SaleContractController {
      */
     @DeleteMapping("/deleteById/{id}")
     public HttpResult deleteByIdApi(@PathVariable("id") Integer id) {
+        //判断是否有物流，如果有不允许删除
+
+        //根据id拿到销售单号
+        QueryWrapper<SaleContract> saleContractQueryWrapper = new QueryWrapper<>();
+        saleContractQueryWrapper.lambda().eq(SaleContract::getId, id);
+        SaleContract one = saleContractService.getOne(saleContractQueryWrapper);
+        String saleContractNo = one.getSaleContractNo();
+        QueryWrapper<LogisticsContract> logisticsContractQueryWrapper = new QueryWrapper<>();
+        logisticsContractQueryWrapper.lambda().eq(LogisticsContract::getSaleContractNo, saleContractNo);
+        List<LogisticsContract> list = logisticsContractService.list(logisticsContractQueryWrapper);
+        if (list != null) {
+            return ResultUtils.error("该销售单已经有物流信息，不允许删除");
+        }
         if (saleContractService.removeById(id)) {
             return ResultUtils.success("删除成功");
         } else {
@@ -94,25 +113,27 @@ public class SaleContractController {
 
     /**
      * 获取当前销售单下的已出库量
+     *
      * @param saleContractNo
      * @return
      */
     @GetMapping("/getRemainingOutboundVolume/{saleContractNo}")
     public HttpResult getRemainingOutboundVolume(@PathVariable("saleContractNo") String saleContractNo) {
         BigDecimal result = saleContractService.getRemainingOutboundVolume(saleContractNo);
-        return ResultUtils.success("查询成功",result);
+        return ResultUtils.success("查询成功", result);
     }
 
 
     /**
      * 根据销售合同编号查到所有的物流详情表
+     *
      * @param saleContractNo
      * @return
      */
     @GetMapping("/getDetailSaleContract/{saleContractNo}")
     public HttpResult getDetailSaleContract(@PathVariable("saleContractNo") String saleContractNo) {
         //根据销售合同找到所有的物流单，再根据物流合同找到所有的物流详情表
-        return ResultUtils.success("查询成功",saleContractService.getDetailSaleContract(saleContractNo));
+        return ResultUtils.success("查询成功", saleContractService.getDetailSaleContract(saleContractNo));
     }
